agents/gcp-adk-a2a/requirements.txt
google_adk
fastapi
a2a-sdk
uvicorn[standard]